Emote handler renaming
